Added to http header 404 response code when not existing page occurs

+++ backend/controllers/PortalController.php
@@ -13,7 +13,6 @@ use common\components\Alert;
 use Exception;
 use Yii;
 use yii\filters\VerbFilter;
-use yii\helpers\Url;
 use yii\web\NotFoundHttpException;
 
 /**
@@ -93,18 +92,22 @@ class PortalController extends BaseController
                     $this->loadSnippetVarValues($portalValueData, $model->portalVarValues[$index]->valueBlock);
                 }
 
-                foreach($model->portalVarValues as $indexPortalVarValue => $portalVarValue) {
+                foreach ($model->portalVarValues as $indexPortalVarValue => $portalVarValue) {
                     $portalVarValue->portal_id = $model->id;
 
                     if ($portalVarValue->removed) {
-                        $portalVarValue->valueBlock->delete();
+
+                        if ($portalVarValue->valueBlock) {
+                            $portalVarValue->valueBlock->delete();
+                        }
                         $portalVarValue->delete();
                         unset($model->portalVarValues[$indexPortalVarValue]);
                         continue;
                     }
 
-                    if (!($portalVarValue->validate() && $portalVarValue->save()))
+                    if (!($portalVarValue->validate() && $portalVarValue->save())) {
                         throw new Exception;
+                    }
 
                     if ($portalVarValue->valueBlock) {
                         if (!($portalVarValue->valueBlock->validate() && $portalVarValue->valueBlock->save())) {
@@ -166,7 +169,7 @@ class PortalController extends BaseController
      */
     public function actionDelete($id)
     {
-        if($this->findModel($id)->delete()){
+        if ($this->findModel($id)->delete()) {
             Alert::success('Položka bola úspešne vymazaná.');
         } else {
             Alert::danger('Vyskytla sa chyba pri vymazávaní položky.');
